OBJFile now use group instead of it own list. Change the group function to find group by name, and now name group when needed, and change tests to use group name instead of number.
Also now the default group is the base group, not a separate group.

Bonus: Add a destructor to cleanup memory
Added a function to get base group. OBJFile will still behave as a valid shape, but now we can skip it in the generated world by adding the base group instead.

Code is now cleaner.